Dashboard: add launching state and split the launch modal into views
Add the pre-launch launching state (a centered spinner under a "Launching
site…" title, sized to hold the modal steady). Extract pure presentational
PreLaunchModal and CelebrationModal views so Storybook can render the real
components without the app's query, router, and analytics providers, and give
each variant its own colocated story with meaningful controls.
